frontendapi/requirements.txt
django
djangorestframework
pika
psycopg2-binary
python-dotenv
dj-database-url